Move TorNode to root package
