+++ network/tor-local-network/src/integrationTest/java/bisq/tor/local_network/DirectoryAuthorityTests.java
@@ -19,7 +19,6 @@ package bisq.tor.local_network;
 
 import bisq.common.util.NetworkUtils;
 import bisq.tor.local_network.da.DirectoryAuthorityFactory;
-import bisq.tor.local_network.da.TorNode;
 import bisq.tor.local_network.torrc.DirectoryAuthorityTorrcGenerator;
 import bisq.tor.local_network.torrc.TorrcFileGenerator;
 import org.junit.jupiter.api.Test;
