+++ network/tor-local-network/src/main/java/bisq/tor/local_network/torrc/RelayTorrcGenerator.java
@@ -17,7 +17,7 @@
 
 package bisq.tor.local_network.torrc;
 
-import bisq.tor.local_network.da.TorNode;
+import bisq.tor.local_network.TorNode;
 
 public class RelayTorrcGenerator extends CommonTorrcGenerator {
     public RelayTorrcGenerator(TorNode thisTorNode) {
