+++ network/tor-local-network/src/test/java/bisq/tor/local_network/DirectoryAuthorityTorrcGeneratorTests.java
@@ -17,7 +17,6 @@
 
 package bisq.tor.local_network;
 
-import bisq.tor.local_network.da.TorNode;
 import bisq.tor.local_network.torrc.DirectoryAuthorityTorrcGenerator;
 import bisq.tor.local_network.torrc.TorrcFileGenerator;
 import org.junit.jupiter.api.Test;
